% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rate.R
\name{check}
\alias{check}
\alias{check_visa}
\title{Check Credit Card Currency Exchange Rate}
\usage{
check_visa(home_cur = "USD", foriegn_cur = "CAD", fee = 0,
  date = today())
}
\description{
Check Credit Card Currency Exchange Rate
}

